+++ tests/test_demography.py
@@ -48,6 +48,94 @@ class TestBadDemographicParameters(unittest.TestCase):
                 ValueError, msprime.PopulationConfiguration, sample_size=bad_size)
 
 
+class TestDeprecatedParameters(unittest.TestCase):
+    """
+    Tests to check that aliased parameters are handled correctly.
+    """
+    def test_mass_migration_dest(self):
+        self.assertRaises(
+            ValueError, msprime.MassMigration, time=0, source=0, dest=0, destination=0)
+        for j in range(10):
+            e = msprime.MassMigration(time=0.1, source=0, dest=j, proportion=0.5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.source, 0)
+            self.assertEqual(e.dest, j)
+            self.assertEqual(e.proportion, 0.5)
+            e = msprime.MassMigration(0.1, 0, j, 0.5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.source, 0)
+            self.assertEqual(e.dest, j)
+            self.assertEqual(e.proportion, 0.5)
+            e = msprime.MassMigration(time=0.1, source=0, destination=j, proportion=0.5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.source, 0)
+            self.assertEqual(e.dest, j)
+            self.assertEqual(e.proportion, 0.5)
+
+    def test_population_parameters_population_id(self):
+        self.assertRaises(
+            ValueError, msprime.PopulationParametersChange,
+            time=0, initial_size=1.1, growth_rate=0.1, population=1,
+            population_id=1)
+
+        for j in range(10):
+            e = msprime.PopulationParametersChange(
+                time=0.1, initial_size=1.1, growth_rate=0.1, population=j)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.initial_size, 1.1)
+            self.assertEqual(e.growth_rate, 0.1)
+            self.assertEqual(e.population, j)
+            e = msprime.PopulationParametersChange(0.1, 1.1, 0.1, j)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.initial_size, 1.1)
+            self.assertEqual(e.growth_rate, 0.1)
+            self.assertEqual(e.population, j)
+            e = msprime.PopulationParametersChange(
+                time=0.1, initial_size=1.1, growth_rate=0.1, population_id=j)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.initial_size, 1.1)
+            self.assertEqual(e.growth_rate, 0.1)
+            self.assertEqual(e.population, j)
+
+    def test_simple_bottleneck_population_id(self):
+        self.assertRaises(
+            ValueError, msprime.SimpleBottleneck,
+            time=0, population=1, proportion=5, population_id=1)
+
+        for j in range(10):
+            e = msprime.SimpleBottleneck(time=0.1, population=j, proportion=5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.population, j)
+            self.assertEqual(e.proportion, 5)
+            e = msprime.SimpleBottleneck(0.1, j, 5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.population, j)
+            self.assertEqual(e.proportion, 5)
+            e = msprime.SimpleBottleneck(time=0.1, population_id=j, proportion=5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.population, j)
+            self.assertEqual(e.proportion, 5)
+
+    def test_instantaneous_bottleneck_population_id(self):
+        self.assertRaises(
+            ValueError, msprime.InstantaneousBottleneck,
+            time=0, population=1, strength=1, population_id=1)
+
+        for j in range(10):
+            e = msprime.InstantaneousBottleneck(time=0.1, population=j, strength=5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.population, j)
+            self.assertEqual(e.strength, 5)
+            e = msprime.InstantaneousBottleneck(0.1, j, 5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.population, j)
+            self.assertEqual(e.strength, 5)
+            e = msprime.InstantaneousBottleneck(time=0.1, population_id=j, strength=5)
+            self.assertEqual(e.time, 0.1)
+            self.assertEqual(e.population, j)
+            self.assertEqual(e.strength, 5)
+
+
 class TestGrowthRates(unittest.TestCase):
     """
     Tests to see the growth rates we calculate give us the
@@ -184,7 +272,7 @@ class TestRateConversions(unittest.TestCase):
         ll_event = {
             "type": "population_parameters_change",
             "time": g,
-            "population_id": -1,
+            "population": -1,
             "initial_size": new_size
         }
         self.assertEqual(event.get_ll_representation(1), ll_event)
@@ -193,11 +281,11 @@ class TestRateConversions(unittest.TestCase):
         g = 512
         growth_rate = 1
         event = msprime.PopulationParametersChange(
-            time=g, growth_rate=growth_rate, population_id=1)
+            time=g, growth_rate=growth_rate, population=1)
         ll_event = {
             "type": "population_parameters_change",
             "time": g,
-            "population_id": 1,
+            "population": 1,
             "growth_rate": growth_rate
         }
         self.assertEqual(event.get_ll_representation(1), ll_event)
@@ -208,11 +296,11 @@ class TestRateConversions(unittest.TestCase):
         initial_size = 8192
         event = msprime.PopulationParametersChange(
             time=g, initial_size=initial_size,
-            growth_rate=growth_rate, population_id=1)
+            growth_rate=growth_rate, population=1)
         ll_event = {
             "type": "population_parameters_change",
             "time": g,
-            "population_id": 1,
+            "population": 1,
             "initial_size": initial_size,
             "growth_rate": growth_rate
         }
@@ -279,11 +367,11 @@ class TestDemographyDebugger(unittest.TestCase):
         demographic_events = [
             msprime.PopulationParametersChange(0.1, initial_size=2),
             msprime.PopulationParametersChange(0.1, growth_rate=10),
-            msprime.MassMigration(0.2, source=1, destination=0),
+            msprime.MassMigration(0.2, source=1, dest=0),
             msprime.MigrationRateChange(0.2, rate=0),
             msprime.MigrationRateChange(0.4, matrix_index=(0, 1), rate=1),
             msprime.MigrationRateChange(0.4, matrix_index=(1, 0), rate=1),
-            msprime.InstantaneousBottleneck(0.5, strength=100)]
+            msprime.InstantaneousBottleneck(0.5, population=0, strength=100)]
         self.verify_debug(
             population_configurations, migration_matrix, demographic_events)
 
@@ -301,8 +389,8 @@ class TestCoalescenceLocations(unittest.TestCase):
         ]
         t = 5
         demographic_events = [
-            msprime.MassMigration(time=t, source=0, destination=2),
-            msprime.MassMigration(time=t, source=1, destination=2),
+            msprime.MassMigration(time=t, source=0, dest=2),
+            msprime.MassMigration(time=t, source=1, dest=2),
         ]
         ts = msprime.simulate(
             population_configurations=population_configurations,
@@ -331,8 +419,8 @@ class TestCoalescenceLocations(unittest.TestCase):
             msprime.PopulationConfiguration(0),
         ]
         demographic_events = [
-            msprime.MassMigration(time=t, source=0, destination=2),
-            msprime.MassMigration(time=t, source=1, destination=2),
+            msprime.MassMigration(time=t, source=0, dest=2),
+            msprime.MassMigration(time=t, source=1, dest=2),
         ]
         ts = msprime.simulate(
             population_configurations=population_configurations,
@@ -405,9 +493,9 @@ class TestCoalescenceLocations(unittest.TestCase):
         ]
         # We migrate the lineages to the next step by step.
         demographic_events = [
-            msprime.MassMigration(time=t1, source=0, destination=1),
-            msprime.MassMigration(time=t2, source=1, destination=2),
-            msprime.MassMigration(time=t3, source=2, destination=3),
+            msprime.MassMigration(time=t1, source=0, dest=1),
+            msprime.MassMigration(time=t2, source=1, dest=2),
+            msprime.MassMigration(time=t3, source=2, dest=3),
         ]
         ts = msprime.simulate(
             population_configurations=population_configurations,
@@ -448,9 +536,9 @@ class TestCoalescenceLocations(unittest.TestCase):
         ]
         # We migrate the lineages to the next step by step.
         demographic_events = [
-            msprime.MassMigration(time=t1, source=0, destination=1),
-            msprime.MassMigration(time=t2, source=1, destination=2),
-            msprime.MassMigration(time=t3, source=2, destination=3),
+            msprime.MassMigration(time=t1, source=0, dest=1),
+            msprime.MassMigration(time=t2, source=1, dest=2),
+            msprime.MassMigration(time=t3, source=2, dest=3),
         ]
         ts = msprime.simulate(
             population_configurations=population_configurations,
@@ -508,7 +596,7 @@ class TestCoalescenceLocations(unittest.TestCase):
             ] + [msprime.PopulationConfiguration(1)]
         t = 5
         demographic_events = [
-            msprime.MassMigration(time=t, source=0, destination=num_demes - 1),
+            msprime.MassMigration(time=t, source=0, dest=num_demes - 1),
         ]
         ts = msprime.simulate(
             population_configurations=population_configurations,
@@ -535,11 +623,11 @@ class TestCoalescenceLocations(unittest.TestCase):
         t3 = 0.0003
         t4 = 0.0004
         demographic_events = [
-            msprime.InstantaneousBottleneck(time=t1, population_id=0, strength=strength),
-            msprime.InstantaneousBottleneck(time=t2, population_id=1, strength=strength),
-            msprime.InstantaneousBottleneck(time=t3, population_id=2, strength=strength),
-            msprime.MassMigration(time=t4, source=2, destination=0),
-            msprime.MassMigration(time=t4, source=1, destination=0)
+            msprime.InstantaneousBottleneck(time=t1, population=0, strength=strength),
+            msprime.InstantaneousBottleneck(time=t2, population=1, strength=strength),
+            msprime.InstantaneousBottleneck(time=t3, population=2, strength=strength),
+            msprime.MassMigration(time=t4, source=2, dest=0),
+            msprime.MassMigration(time=t4, source=1, dest=0)
         ]
         ts = msprime.simulate(
             population_configurations=population_configurations,
@@ -551,7 +639,7 @@ class TestCoalescenceLocations(unittest.TestCase):
         # The parent of all the samples from each deme should be in that deme.
         for pop in range(3):
             parents = [
-                tree.get_parent(u) for u in ts.get_samples(population_id=pop)]
+                tree.get_parent(u) for u in ts.samples(population=pop)]
             for v in parents:
                 self.assertEqual(tree.get_population(v), pop)
 
@@ -568,8 +656,8 @@ class TestMigrationRecords(unittest.TestCase):
         ]
         t = 5
         demographic_events = [
-            msprime.MassMigration(time=t, source=0, destination=2),
-            msprime.MassMigration(time=t, source=1, destination=2),
+            msprime.MassMigration(time=t, source=0, dest=2),
+            msprime.MassMigration(time=t, source=1, dest=2),
         ]
         ts = msprime.simulate(
             population_configurations=population_configurations,
@@ -608,7 +696,7 @@ class TestTimeUnits(unittest.TestCase):
         # have a very strong negative growth rate, resulting in almost instant
         # coalescence.
         demographic_events = [
-            msprime.MassMigration(time=g, source=1, destination=0),
+            msprime.MassMigration(time=g, source=1, dest=0),
             msprime.PopulationParametersChange(time=g, growth_rate=1000),
         ]
         reps = msprime.simulate(
@@ -638,7 +726,7 @@ class TestTimeUnits(unittest.TestCase):
         # have a very strong bottleneck, resulting in almost instant
         # coalescence.
         demographic_events = [
-            msprime.MassMigration(time=g, source=1, destination=0),
+            msprime.MassMigration(time=g, source=1, dest=0),
             msprime.PopulationParametersChange(time=g, initial_size=1e-3),
         ]
         reps = msprime.simulate(
@@ -664,8 +752,8 @@ class TestTimeUnits(unittest.TestCase):
         # have a very strong bottleneck, resulting in instant
         # coalescence.
         demographic_events = [
-            msprime.MassMigration(time=t, source=1, destination=0),
-            msprime.InstantaneousBottleneck(time=t, strength=100)
+            msprime.MassMigration(time=t, source=1, dest=0),
+            msprime.InstantaneousBottleneck(time=t, population=0, strength=100)
         ]
         reps = msprime.simulate(
             Ne=Ne,
@@ -722,7 +810,7 @@ class TestLowLevelConversions(unittest.TestCase):
                 d = event.get_ll_representation(1)
                 dp = {
                     "time": g,
-                    "population_id": -1,
+                    "population": -1,
                     "type": "population_parameters_change",
                     "initial_size": Ne}
                 self.assertEqual(d, dp)
@@ -736,7 +824,7 @@ class TestLowLevelConversions(unittest.TestCase):
                 d = event.get_ll_representation(1)
                 dp = {
                     "time": g,
-                    "population_id": -1,
+                    "population": -1,
                     "type": "population_parameters_change",
                     "initial_size": initial_size}
                 self.assertEqual(d, dp)
@@ -748,21 +836,21 @@ class TestLowLevelConversions(unittest.TestCase):
             d = event.get_ll_representation(1)
             dp = {
                 "time": g,
-                "population_id": -1,
+                "population": -1,
                 "type": "population_parameters_change",
                 "growth_rate": growth_rate}
             self.assertEqual(d, dp)
 
-    def test_population_parameters_change_population_id(self):
+    def test_population_parameters_change_population(self):
         g = 100
         Ne = 10
-        for population_id in range(3):
+        for population in range(3):
             event = msprime.PopulationParametersChange(
-                time=g, initial_size=Ne, population_id=population_id)
+                time=g, initial_size=Ne, population=population)
             d = event.get_ll_representation(1)
             dp = {
                 "time": g,
-                "population_id": population_id,
+                "population": population,
                 "type": "population_parameters_change",
                 "initial_size": Ne}
             self.assertEqual(d, dp)
@@ -805,39 +893,39 @@ class TestLowLevelConversions(unittest.TestCase):
 
     def test_mass_migration_time(self):
         for g in [0.1, 1, 100, 1e6]:
-            event = msprime.MassMigration(time=g, source=0, destination=1)
+            event = msprime.MassMigration(time=g, source=0, dest=1)
             d = event.get_ll_representation(1)
             dp = {
                 "time": g,
                 "type": "mass_migration",
                 "source": 0,
-                "destination": 1,
+                "dest": 1,
                 "proportion": 1}
             self.assertEqual(d, dp)
 
     def test_mass_migration_source_dest(self):
         g = 51
         for source, dest in itertools.permutations(range(4), 2):
-            event = msprime.MassMigration(time=g, source=source, destination=dest)
+            event = msprime.MassMigration(time=g, source=source, dest=dest)
             d = event.get_ll_representation(1)
             dp = {
                 "time": g,
                 "type": "mass_migration",
                 "source": source,
-                "destination": dest,
+                "dest": dest,
                 "proportion": 1}
             self.assertEqual(d, dp)
 
     def test_mass_migration_proportion(self):
         g = 51
         for p in [0, 1e-6, 0.4, 1]:
-            event = msprime.MassMigration(time=g, source=0, destination=1, proportion=p)
+            event = msprime.MassMigration(time=g, source=0, dest=1, proportion=p)
             d = event.get_ll_representation(1)
             dp = {
                 "time": g,
                 "type": "mass_migration",
                 "source": 0,
-                "destination": 1,
+                "dest": 1,
                 "proportion": p}
             self.assertEqual(d, dp)
 
@@ -859,12 +947,12 @@ class TestLowLevelConversions(unittest.TestCase):
         for population in [0, 1, 5]:
             for strength in [0, 100, 1000, 1e9]:
                 event = msprime.InstantaneousBottleneck(
-                    time=g, population_id=population, strength=strength)
+                    time=g, population=population, strength=strength)
                 d = event.get_ll_representation(1)
                 dp = {
                     "time": g,
                     "type": "instantaneous_bottleneck",
-                    "population_id": population,
+                    "population": population,
                     "strength": strength}
                 self.assertEqual(d, dp)
 
@@ -935,7 +1023,7 @@ class TestHistoricalSampling(unittest.TestCase):
                 msprime.PopulationConfiguration()],
             demographic_events=[
                 msprime.MassMigration(
-                    time=migration_time, source=1, destination=0)])
+                    time=migration_time, source=1, dest=0)])
         t = next(ts.trees())
         self.assertEqual(t.get_time(0), 0)
         self.assertEqual(t.get_time(1), sampling_time)
@@ -961,9 +1049,9 @@ class TestHistoricalSampling(unittest.TestCase):
                 msprime.PopulationConfiguration(),
                 msprime.PopulationConfiguration()],
             demographic_events=[
-                msprime.MassMigration(time=t1, source=0, destination=1),
-                msprime.MassMigration(time=t2, source=1, destination=2),
-                msprime.MassMigration(time=t3, source=2, destination=3)])
+                msprime.MassMigration(time=t1, source=0, dest=1),
+                msprime.MassMigration(time=t2, source=1, dest=2),
+                msprime.MassMigration(time=t3, source=2, dest=3)])
         t = next(ts.trees())
         self.assertEqual(t.get_time(0), 0)
         self.assertEqual(t.get_time(1), t1)